+++ examples/src/stories/basic.tsx
@@ -18,7 +18,6 @@
 import { Flow, Location } from '@flowmap.gl/core';
 import FlowMap, { DiffColorsLegend, getViewStateForFeatures, LegendBox, LocationTotalsLegend } from '@flowmap.gl/react';
 import { storiesOf } from '@storybook/react';
-import { quantile } from 'd3-array';
 import * as d3scaleChromatic from 'd3-scale-chromatic';
 import React from 'react';
 import NonInteractiveExample from '../components/NonInteractiveExample';
@@ -50,31 +49,6 @@ storiesOf('Basic', module)
         </LegendBox>
       </>
     )),
-  )
-  .add(
-    '75 percentile pickable',
-    pipe(
-      withStats,
-      withFetchJson('locations', './data/locations.json'),
-      withFetchJson('flows', './data/flows-2016.json'),
-    )(({ locations, flows }: any) => {
-      const q = quantile(flows, 0.75, (f: Flow) => Math.abs(f.count));
-      return (
-        <>
-          <FlowMap
-            getLocationId={getLocationId}
-            getFlowPickable={q != null ? (f: Flow) => f.count >= q : undefined}
-            flows={flows}
-            locations={locations}
-            initialViewState={getViewStateForFeatures(locations, [window.innerWidth, window.innerHeight])}
-            mapboxAccessToken={mapboxAccessToken}
-          />
-          <LegendBox bottom={35} left={10}>
-            <LocationTotalsLegend />
-          </LegendBox>
-        </>
-      );
-    }),
   )
   .add(
     'custom flow color scheme',
@@ -399,6 +373,55 @@ storiesOf('Basic', module)
             maxFlowThickness={15}
             flows={flows}
             animate={true}
+            locations={locations}
+            initialViewState={getViewStateForFeatures(locations, [window.innerWidth, window.innerHeight])}
+            mapboxAccessToken={mapboxAccessToken}
+          />
+          <LegendBox bottom={35} left={10}>
+            <LocationTotalsLegend />
+          </LegendBox>
+        </>
+      );
+    }),
+  )
+  .add(
+    'minPickableFlowThickness',
+    pipe(
+      withStats,
+      withFetchJson('locations', './data/locations.json'),
+      withFetchJson('flows', './data/flows-2016.json'),
+    )(({ locations, flows }: any) => {
+      return (
+        <>
+          <FlowMap
+            getLocationId={getLocationId}
+            minPickableFlowThickness={1}
+            flows={flows}
+            locations={locations}
+            initialViewState={getViewStateForFeatures(locations, [window.innerWidth, window.innerHeight])}
+            mapboxAccessToken={mapboxAccessToken}
+          />
+          <LegendBox bottom={35} left={10}>
+            <LocationTotalsLegend />
+          </LegendBox>
+        </>
+      );
+    }),
+  )
+  .add(
+    'minPickableFlowThickness animated',
+    pipe(
+      withStats,
+      withFetchJson('locations', './data/locations.json'),
+      withFetchJson('flows', './data/flows-2016.json'),
+    )(({ locations, flows }: any) => {
+      return (
+        <>
+          <FlowMap
+            getLocationId={getLocationId}
+            minPickableFlowThickness={1}
+            animate={true}
+            flows={flows}
             locations={locations}
             initialViewState={getViewStateForFeatures(locations, [window.innerWidth, window.innerHeight])}
             mapboxAccessToken={mapboxAccessToken}