+++ packages/core/src/AnimatedFlowLinesLayer/AnimatedFlowLinesLayerVertex.glsl.ts
@@ -52,6 +52,7 @@ attribute vec4 instanceColors;
 attribute vec3 instancePickingColors;
 attribute float instanceWidths;
 attribute float instanceStaggering;
+attribute float instancePickable;
          
 uniform float opacity;
 uniform float currentTime;
@@ -92,6 +93,8 @@ void main(void) {
   sourceToTarget = positions.x * length(source - target) * NUM_PARTS - currentTime * SPEED + instanceStaggering; 
 
   // Set color to be rendered to picking fbo (also used to check for selection highlight).
-  picking_setPickingColor(instancePickingColors);
+  if (instancePickable > 0.5) {
+    picking_setPickingColor(instancePickingColors);
+  }
 }
 `;