+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 - Yarn for building
 - Upgraded deck.gl
+- Improved highlighted location area coloring; drawing location areas outline above arrows
 - Added staggering to animated flow lines
 - Added pickingOnlyTopFlows for specifying the number of top flows which are pickable 
 - Added highlightedLocationAreaId for pre-hovering