+++ CHANGELOG.md
@@ -6,8 +6,9 @@
 - Upgraded deck.gl
 - Improved highlighted location area coloring; drawing location areas outline above arrows
 - Added staggering to animated flow lines
-- Added getFlowPickable 
 - Added highlightedLocationAreaId for pre-hovering
+- Added maxFlowThickness
+- Added minPickableFlowThickness 
 
 
 ## [6.0.0] - 2019-08-20